# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AED Foreigner API",
        "description": "API documentation for AED Foreigner Enrollment",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/login": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Admin/Agent Login (Step 1: Request OTP)",
                "operationId": "adminLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent to email"
                    },
                    "403": {
                        "description": "Forbidden (Not an agent)"
                    }
                }
            }
        },
        "/api/v1/agents/{id}": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Update Agent Details",
                "operationId": "updateAgent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "LEVEL1",
                                            "LEVEL2",
                                            "LEVEL3",
                                            "SUPERVISEUR",
                                            "AUDITEUR"
                                        ]
                                    },
                                    "phonenumber": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agent updated"
                    }
                }
            }
        },
        "/api/v1/admins/logout": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Admin Logout",
                "operationId": "adminLogout",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/agents/register": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Register New Agent",
                "operationId": "registerAgent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phonenumber",
                                    "npi",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string"
                                    },
                                    "phonenumber": {
                                        "type": "string"
                                    },
                                    "npi": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agent registered"
                    }
                }
            }
        },
        "/api/v1/admins/verify-otp": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Admin/Agent Verify OTP (Step 2: Get Token)",
                "operationId": "adminVerifyOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "otp": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token obtained"
                    },
                    "403": {
                        "description": "Forbidden or Invalid OTP"
                    }
                }
            }
        },
...
    },
...
}